feat(auth): Add auth to tasks routes

+++ src/controllers/tasks.controller.ts
@@ -1,19 +1,29 @@
+import type { Task } from "../entities/Task.ts";
+import { HTTP_STATUS } from "../modules/router/const.ts";
 import {
-  HTTP_STATUS,
   type RequestContext,
   type RequestMod,
-} from "../modules/router/index.ts";
-import { type ResponseMod } from "../modules/router/index.ts";
-
+  type ResponseMod,
+} from "../modules/router/types.ts";
 import * as repository from "../repositories/task.repository.ts";
 import type {
   CreateTaskPayload,
   UpdateTaskPayload,
 } from "../repositories/task.repository.ts";
-import { handleError } from "../utils/http.ts";
+import { handleError, ResError } from "../utils/http.ts";
 
-export const getTasks = (_req: RequestMod, res: ResponseMod) => {
-  repository.getTasks((err, tasks) => {
+export type TaskDto = Omit<Task, "userId">;
+
+export const convertToDto = (task: Task): TaskDto => {
+  const { userId, ...taskDto } = task;
+
+  return taskDto;
+};
+
+const getTasks = (req: RequestMod, res: ResponseMod) => {
+  const userId = req.context.user?.id as string;
+
+  repository.getTasksByUserId(userId, (err, tasks) => {
     if (err) {
       return handleError(res, err);
     }
@@ -21,12 +31,13 @@ export const getTasks = (_req: RequestMod, res: ResponseMod) => {
     res.writeHead(HTTP_STATUS.success, {
       "Content-Type": "application/json",
     });
-    res.end(JSON.stringify({ tasks: tasks }));
+
+    res.end(JSON.stringify({ tasks: tasks.map(convertToDto) }));
   });
 };
 
-export const getTask = (
-  _req: RequestMod,
+const getTask = (
+  req: RequestMod,
   res: ResponseMod,
   { params }: RequestContext,
 ) => {
@@ -35,64 +46,116 @@ export const getTask = (
       return handleError(res, err);
     }
 
+    if (task.userId !== req.context.user?.id) {
+      res.writeHead(HTTP_STATUS.forbidden, {
+        "Content-Type": "application/json",
+      });
+      res.end(JSON.stringify({ message: "Forbidden" }));
+
+      return;
+    }
+
+    if (!task) {
+      return handleError(res, new ResError({ cause: "not-found" }));
+    }
+
     res.writeHead(HTTP_STATUS.success, {
       "Content-Type": "application/json",
     });
-    res.end(JSON.stringify({ task: task }));
+    res.end(JSON.stringify({ task: convertToDto(task) }));
   });
 };
 
-export const createTask = (
-  _req: RequestMod,
+const createTask = (
+  req: RequestMod,
   res: ResponseMod,
   { body }: RequestContext,
 ) => {
+  const userId = req.context.user?.id as string;
   const payload = body as CreateTaskPayload;
 
-  repository.createTask(payload, (err, task) => {
+  repository.createTask(userId, payload, (err, task) => {
     if (err) {
       return handleError(res, err);
+    }
+
+    if (!task) {
+      return handleError(res, new ResError({ cause: "not-found" }));
     }
 
     res.writeHead(HTTP_STATUS.created, {
       "Content-Type": "application/json",
     });
-    res.end(JSON.stringify({ task: task }));
+    res.end(JSON.stringify({ task: convertToDto(task) }));
   });
 };
 
-export const updateTask = (
-  _req: RequestMod,
+const updateTask = (
+  req: RequestMod,
   res: ResponseMod,
   { params, body }: RequestContext,
 ) => {
   const payload = body as UpdateTaskPayload;
 
-  repository.updateTask(params.id, payload, (err, task) => {
+  repository.getTask(params.id, (err, task) => {
     if (err) {
       return handleError(res, err);
     }
 
-    res.writeHead(HTTP_STATUS.success, {
-      "Content-Type": "application/json",
+    if (task.userId !== req.context.user?.id) {
+      handleError(res, new Error("", { cause: "forbidden" }));
+      return;
+    }
+
+    repository.updateTask(params.id, payload, (err, task) => {
+      if (err) {
+        return handleError(res, err);
+      }
+
+      if (!task) {
+        return handleError(res, new ResError({ cause: "not-found" }));
+      }
+
+      res.writeHead(HTTP_STATUS.success, {
+        "Content-Type": "application/json",
+      });
+      res.end(JSON.stringify({ task: convertToDto(task) }));
     });
-    res.end(JSON.stringify({ task }));
   });
 };
 
-export const deleteTask = (
-  _req: RequestMod,
+const deleteTask = (
+  req: RequestMod,
   res: ResponseMod,
   { params }: RequestContext,
 ) => {
-  repository.deleteTask(params.id, (err) => {
+  repository.getTask(params.id, (err, task) => {
     if (err) {
       return handleError(res, err);
     }
 
-    res.writeHead(HTTP_STATUS.success, {
-      "Content-Type": "application/json",
+    if (task.userId !== req.context.user?.id) {
+      handleError(res, new ResError({ cause: "forbidden" }));
+      return;
+    }
+
+    repository.deleteTask(params.id, (err) => {
+      if (err) {
+        return handleError(res, err);
+      }
+
+      res.writeHead(HTTP_STATUS.success, {
+        "Content-Type": "application/json",
+      });
+      res.end(JSON.stringify({ message: "Deleted" }));
     });
-    res.end(JSON.stringify({ message: "Deleted" }));
   });
+};
+
+export const tasksController = {
+  getTasks,
+  getTask,
+  createTask,
+  updateTask,
+  deleteTask,
 };
